remove mobile-menu.tsx
